Improve handshake popup
